Moves date time extension to extension module
The `ToWeekday` trait and its implementation for `chrono::DateTime`
were previously defined in `expression.rs`. This commit moves them to
a new `date_time_ext.rs` file within the `extensions` module. This
change promotes code reusability and maintains a cleaner structure.
