+++ src/extensions/mod.rs
@@ -1,2 +1,3 @@
+pub mod date_time_ext;
 pub mod path_ext;
 pub mod unsigned_ints_ext;
